fix: error type

+++ sources/ModCore/Events/Interfaces/Game/IOnFrameUpdate.cs
@@ -3,6 +3,6 @@
     [Event]
     public interface IOnFrameUpdate
     {
-        public void OnFrameUpdate( float dt );
+        public void OnFrameUpdate( double dt );
     }
 }
